+++ assets/src/scripts/components/map/index.js
@@ -8,8 +8,10 @@ import { FLOOD_EXTENTS_ENDPOINT, FLOOD_BREACH_ENDPOINT, FLOOD_LEVEE_ENDPOINT } f
 import { hasFloodData, getFloodExtent, getFloodStageHeight } from '../../selectors/flood-data-selector';
 import { Actions } from '../../store';
 import { floodSlider } from './flood-slider';
-import { createLegendControl, createFIMLegend } from './legend';
-import { addNldi } from './nldiMapping';
+import { createLegendControl, createFIMLegend, createNldiLegend } from './legend';
+import { addNldiLayers} from './nldiMapping';
+import { hasNldiData, getNldiDownstreamFlows, getNldiDownstreamSites, getNldiUpstreamFlows, getNldiUpstreamSites}
+    from '../../selectors/nldi-data-selector';
 
 
 const getLayerDefs = function(layerNo, siteno, stage) {
@@ -93,6 +95,11 @@ const siteMap = function(node, {siteno, latitude, longitude, zoom}) {
     };
 
 
+    const updateNldiLayers = function (node, {upstreamFlows, downstreamFlows, upstreamSites, downstreamSites}) {
+        addNldiLayers(map, upstreamFlows, downstreamFlows, upstreamSites, downstreamSites);
+    };
+
+
     const updateMapExtent = function (node, extent) {
         if (Object.keys(extent).length > 0) {
             map.fitBounds(Util.extentToBounds(extent).extend([latitude, longitude]));
@@ -141,8 +148,14 @@ const siteMap = function(node, {siteno, latitude, longitude, zoom}) {
     // Add a marker at the site location
     createMarker([latitude, longitude]).addTo(map);
 
-    //add nldi layers
-    addNldi(map, legendControl, siteno);
+    /*
+     * Creates the NLDI legend if NLDI data is available, otherwise removes the NLDI legend if it exists.
+     * @param {HTMLElement} node - element where the map is rendered
+     * @param {Boolean} isNldiAvailable
+     */
+    const addNldiLegend = function(node, isNldiAvailable) {
+        createNldiLegend(legendControl, isNldiAvailable);
+    };
 
     node
         .call(link(updateFloodLayers, createStructuredSelector({
@@ -151,7 +164,14 @@ const siteMap = function(node, {siteno, latitude, longitude, zoom}) {
         })))
         .call(link(updateMapExtent, getFloodExtent))
         .call(link(addFIMLegend, hasFloodData))
-        .call(link(addFimLink, hasFloodData));
+        .call(link(addFimLink, hasFloodData))
+        .call(link(addNldiLegend, hasNldiData))
+        .call(link(updateNldiLayers, createStructuredSelector({
+            upstreamFlows: getNldiUpstreamFlows,
+            downstreamFlows: getNldiDownstreamFlows,
+            upstreamSites: getNldiUpstreamSites,
+            downstreamSites: getNldiDownstreamSites
+        })));
 };
 
 /*
@@ -166,6 +186,8 @@ const siteMap = function(node, {siteno, latitude, longitude, zoom}) {
 export const attachToNode = function(store, node, {siteno, latitude, longitude, zoom}) {
 
     store.dispatch(Actions.retrieveFloodData(siteno));
+    // hydrates the store with nldi data
+    store.dispatch(Actions.retrieveNldiData(siteno));
 
     select(node)
         .call(provide(store));