+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/usgs/waterdataui/compare/waterdataui-0.21.0...master)
+### Changed
+- NLDI data is now loaded into redux store
+- Legend expand/collapse is available when nldi data is present
+- Legend is initially collapsed on small devices when nldi data or flood data is present
 
 ## [0.21.0](https://github.com/usgs/waterdataui/compare/waterdataui-0.20.0...waterdataui-0.21.0) - 2019-10-24
 ### Added